Backend/New_App/requirements.txt
Flask
celery[redis]
redis
flask-sqlalchemy
email-validator
bcrypt
flask_bcrypt
flask-cors
PyJWT
datetime
flask_login
flask_wtf
numpy
python-dotenv
requests
schedule
twilio
Flask-Caching
weasyprint